add locks

+++ server/lyft/gateway/autoplan_builder.go
@@ -44,6 +44,7 @@ type AutoplanValidator struct {
 	PrjCmdBuilder              events.ProjectPlanCommandBuilder
 	PullUpdater                *events.PullUpdater
 	WorkingDir                 events.WorkingDir
+	WorkingDirLocker           events.WorkingDirLocker
 }
 
 func (r *AutoplanValidator) isValid(baseRepo models.Repo, headRepo models.Repo, pull models.PullRequest, user models.User) (bool, error) {
@@ -80,12 +81,24 @@ func (r *AutoplanValidator) isValid(baseRepo models.Repo, headRepo models.Repo, 
 			ctx.Log.Warn("unable to update commit status: %s", statusErr)
 		}
 		// If error happened after clone was made, we should clean it up here too
+		unlockFn, lockErr := r.WorkingDirLocker.TryLock(baseRepo.FullName, pull.Num, "default")
+		if lockErr != nil {
+			ctx.Log.Warn("workspace was locked")
+			return false, errors.Wrap(err, lockErr.Error())
+		}
+		defer unlockFn()
 		if cloneErr := r.WorkingDir.Delete(baseRepo, pull); cloneErr != nil {
 			ctx.Log.With("err", cloneErr).Warn("unable to delete clone after autoplan failed")
 		}
 		r.PullUpdater.UpdatePull(ctx, events.AutoplanCommand{}, command.Result{Error: err})
 		return false, errors.Wrap(err, "Failed building autoplan commands")
 	}
+	unlockFn, err := r.WorkingDirLocker.TryLock(baseRepo.FullName, pull.Num, "default")
+	if err != nil {
+		ctx.Log.Warn("workspace was locked")
+		return false, err
+	}
+	defer unlockFn()
 	// Delete repo clone generated to validate plan
 	if err := r.WorkingDir.Delete(baseRepo, pull); err != nil {
 		return false, errors.Wrap(err, "Failed deleting cloned repo")
